duplicated widgets removed (biomass and alerts)
