+++ src/components/widget/style.module.scss
@@ -148,7 +148,7 @@
   &.mangrove_coverage,
   .mangrove_extent,
   .mangrove_net_change,
-  .mangrove_biomass_pie,
+  .mangrove_biomass,
   .mangrove_blue_carbon {
     margin: 30px 0 0 0;
     height: 90%;
