+++ src/components/widgets/index.js
@@ -12,10 +12,8 @@ import MangroveNetChange from 'widget-components/mangrove-net-change';
 import MangroveBlueCarbon from 'widget-components/mangrove-blue-carbon';
 import MangroveActivity from 'widget-components/mangrove-activity';
 import MangroveAlerts from 'widget-components/mangrove-alerts';
-import MangroveAlertsLine from 'widget-components/mangrove-alerts-line';
 import MangroveHeight from 'widget-components/mangrove-height';
 import MangroveBiomass from 'widget-components/mangrove-biomass';
-import MangroveBiomassPie from 'widget-components/mangrove-biomass-pie';
 import ConservationHotspots from 'widget-components/conservation-hotspots';
 
 import Component from './component';
@@ -45,17 +43,11 @@ const templates = new Map([
   ['mangrove_alerts', {
     component: MangroveAlerts
   }],
-  ['mangrove_alerts_line', {
-    component: MangroveAlertsLine
-  }],
   ['mangrove_height', {
     component: MangroveHeight
   }],
   ['mangrove_biomass', {
     component: MangroveBiomass
-  }],
-  ['mangrove_biomass_pie', {
-    component: MangroveBiomassPie
   }],
   ['conservation_hotspots', {
     component: ConservationHotspots
