--- src/widget-components/mangrove-alerts-line/constants.js
@@ -1,64 +1,0 @@
-const realData = [
-  {
-    date: {
-      value: "2019-07-01"
-    },
-    count: 372
-  },
-  {
-    date: {
-      value: "2019-11-01"
-    },
-    count: 578
-  },
-  {
-    date: {
-      value: "2019-10-01"
-    },
-    count: 1108
-  },
-  {
-    date: {
-      value: "2019-09-01"
-    },
-    count: 526
-  },
-  {
-    date: {
-      value: "2019-03-01"
-    },
-    count: 426
-  },
-  {
-    date: {
-      value: "2019-06-01"
-    },
-    count: 1218
-  },
-  {
-    date: {
-      value: "2019-05-01"
-    },
-    count: 180
-  },
-  {
-    date: {
-      value: "2019-02-01"
-    },
-    count: 154
-  },
-  {
-    date: {
-      value: "2019-08-01"
-    },
-    count: 250
-  },
-  {
-    date: {
-      value: "2019-04-01"
-    },
-    count: 96
-  },
-];
-
-export default realData;
