--- src/widget-components/mangrove-alerts-line/index.js
@@ -1,22 +1,0 @@
-import { connect } from 'react-redux';
-import { fetchAlerts } from 'modules/alerts/actions';
-import { setUi } from 'modules/widgets/actions';
-import Component from './component';
-
-const mapStateToProps = state => ({
-  data: state.alerts,
-  isLoading: state.ranking.isLoading,
-  currentLocation: state.locations.current,
-  locationsList: state.locations.list,
-  ui: state.widgets.ui.alerts || {
-    alerts: {
-      startDate: '2020-04-01',
-      endDate: '2020-05-01',
-      year: 2020
-    }
-  }
-});
-
-const mapDispatchToProps = { setUi, fetchAlerts };
-
-export default connect(mapStateToProps, mapDispatchToProps)(Component);
