--- src/widget-components/mangrove-biomass-pie/index.js
@@ -1,16 +1,0 @@
-import { connect } from 'react-redux';
-import { setUi } from 'modules/widgets/actions';
-
-import Component from './component';
-
-const mapStateToProps = state => ({
-  isLoading: state.mangroveData.isLoading,
-  data: state.mangroveData,
-  ui: state.widgets.ui.coverage || {
-    currentYear: 2016,
-    unit: '%'
-  }
-});
-const mapDispatchToProps = { setUi };
-
-export default connect(mapStateToProps, mapDispatchToProps)(Component);
